+++ src/main/java/ru/topjava/lunchvoter/repository/CrudUserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.topjava.lunchvoter.model.User;
 
 @Transactional(readOnly = true)
-public interface CrudUserRepository extends JpaRepository<User, Integer> {
+public interface CrudUserRepository extends JpaRepository<User, Long> {
 
 }
