+++ src/test/java/ru/topjava/lunchvoter/MenuTestData.java
@@ -7,9 +7,9 @@ import java.time.LocalDate;
 import static ru.topjava.lunchvoter.RestaurantTestData.RESTAURANT_GINZA;
 
 public class MenuTestData {
-    public static final int LUNCH_MENU_SEQ_START = 1;
-    public static final int LUNCH_MENU_GINZA_ID = LUNCH_MENU_SEQ_START;
-    public static final int LUNCH_MENU_MAMA_ITALIA_ID = LUNCH_MENU_SEQ_START + 1;
+    public static final Long LUNCH_MENU_SEQ_START = 1L;
+    public static final Long LUNCH_MENU_GINZA_ID = LUNCH_MENU_SEQ_START;
+    public static final Long LUNCH_MENU_MAMA_ITALIA_ID = LUNCH_MENU_SEQ_START + 1;
     public static LocalDate LUNCH_MENU_DATE = LocalDate.of(2024, 4, 1);
     public static LocalDate NON_EXISTING_MENU_DATE = LocalDate.of(1900, 4, 1);
     public static LocalDate LUNCH_MENU_NEW_DATE = LocalDate.of(2055, 11, 12);
